Rename `codec` -> `codecs`
